stm32mp1: move PSCI in pm/psci.c
Source file pm/psci.c will soon hold several PSCI functions and sequences.
Move now the existing PSCI_CPU_ON support to pm/psci.c.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
