+++ core/arch/arm/plat-stm32mp1/pm/sub.mk
@@ -1,0 +1,1 @@
+srcs-$(CFG_PSCI_ARM32) += psci.c
