+++ core/arch/arm/plat-stm32mp1/sub.mk
@@ -2,3 +2,5 @@ global-incdirs-y += .
 
 srcs-y += main.c
 srcs-y += reset.S
+
+subdirs-y += pm
